Tweak working directory behavior when using chroot or pivot_root
The former behavior was to try and get the current process's working
directory outside the chroot, then try to map it to a new path within
the chroot. If the current working directory was outside the chroot
directory tree, then we would fallback to just using `/` as the working
directory.

Because paths were resolved relative to the chroot _before_ the callback
for `Command::before_chroot` was called, this caused an inconsistency
where it could be impossible to use some directories as the working
directory in the new process (say, when `before_chroot` is used to set
up bind mounts).

The new behavior is simpler: using chroot or pivot_root _always_ changes
to `/` as the working directory, then `Command::current_dir` will be
applied relative to the new root.